requirements_dev.txt
-r requirements.txt
alabaster==0.7.12
argon2-cffi==21.1.0
asgiref==3.4.1
astroid==2.9.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docutils==0.17.1
entrypoints==0.3
idna==3.3
imagesize==1.3.0
ipykernel==6.6.0
ipyparallel==8.0.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
nose==1.3.7
notebook==6.4.6
numpy==1.21.4
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
pylint==2.12.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
requests==2.26.0
Send2Trash==1.8.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wrapt==1.13.3
model-mommy
django-debug-toolbar
django-browser-reload
